Create first version of CSS project "Landing Page"
This commit is the first version of the project with minimal reliance on notes in order to practice recall and cement current learning content. Plan to refactor after some review, as I relied heavily on margin and padding with fixed pixel values, which causes problems when the window is resized.